# Improve unfilter() performance