@HeroicKatora 2023-04-19
Impressive (auto vectorization?) wins, very nicely spotted to add those monomorphizations manually.